Web App/requirements.txt
Flask==2.0.1
cryptography==3.4.7
google-api-python-client
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-cloud-pubsub==2.7.0
google-cloud-iot==2.2.1
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
google-cloud-aiplatform
pyjwt==2.1.0
requests==2.26.0
paho-mqtt==1.5.1
pillow==8.4.0
dash
pandas
numpy
